refactor: unite search statistics in 1 file
