+++ src/storages/mongo/statistics.py
@@ -28,3 +28,19 @@ class SearchStatistics(SearchStatisticsSchema, CustomDocument):
             IndexModel([("query", pymongo.TEXT)], name="query_index"),
             IndexModel([("created_at", pymongo.DESCENDING)], name="created_at_index"),
         ]
+
+
+class AskStatisticsSchema(CustomModel):
+    query: str
+    answer: str
+    search_responses: list[WrappedResponseSchema]
+    time_spent: float  # Time spent on the search in seconds
+    created_at: datetime.datetime = Field(default_factory=lambda: datetime.datetime.now(datetime.UTC))
+
+
+class AskStatistics(AskStatisticsSchema, CustomDocument):
+    class Settings:
+        indexes = [
+            IndexModel([("query", pymongo.TEXT)], name="query_index"),
+            IndexModel([("created_at", pymongo.DESCENDING)], name="created_at_index"),
+        ]
